Split tests into separeate files
