+++ tests/test_data.py
@@ -1,32 +1,3 @@
-from openai import OpenAI
-import psycopg2
-import os
-import time
-import pytest
-import sys
-import dotenv
-from nltk import download, sent_tokenize
-from nltk.langnames import langname
-import math
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../algorithm")))
-dotenv.load_dotenv()
-
-from find_k_nearest import find_k_nearest
-from create_embeddings import create_embeddings, create_embeddings_multithreading
-
-
-def cosine_similarity(v1, v2):
-    "compute cosine similarity of v1 to v2: (v1 dot v2)/{||v1||*||v2||)"
-    sumxx, sumxy, sumyy = 0, 0, 0
-    for i in range(len(v1)):
-        x = v1[i]
-        y = v2[i]
-        sumxx += x * x
-        sumyy += y * y
-        sumxy += x * y
-    return sumxy / math.sqrt(sumxx * sumyy)
-
 
 TEST_TEXT = """
 The cat jumped over the tall wooden fence.
@@ -130,81 +101,3 @@ The door creaked open to reveal a forgotten attic.
 He planted a tiny seedling in the fertile soil.
 The echo of her voice lingered long after she had spoken.
 """
-
-
-def test_find_k_nearest():
-    openai_client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
-    model_name = os.getenv("MODEL_NAME")
-
-    conn = psycopg2.connect(
-        dbname=os.getenv("PGDATABASE"),
-        user=os.getenv("PGUSER"),
-        password=os.getenv("PGPASSWORD"),
-        host=os.getenv("PGHOST"),
-        port=os.getenv("PGPORT", 5432),
-    )
-
-    vec_lenght = int(os.getenv("VECTOR_LENGTH"))
-
-    # This sententeded was translated from Polish into English
-    # We are hoping to find the same sentence but in the original language
-    translated_sent = "2015 saw the premiere of ''The Last Jedi'' (2017) and ''The Rise of Skywalker."
-    org_lang_title = "Gwiezdne wojny"
-    org_lang_doc_index = 11
-    n_articles = 2
-
-    # embeddings for translated sentece
-    response = openai_client.embeddings.create(input=translated_sent, model=model_name)
-    vector = response.data[0].embedding
-
-    # calcualating the distance
-    start_time = time.time()
-    test_output = find_k_nearest(vector, n_articles, conn, vec_lenght, "en")
-    print(f"find_k_nearest execution time {time.time() - start_time}s")
-
-    assert len(test_output) == n_articles
-
-    # testing if desired text was found
-    assert test_output[0][0] == org_lang_title
-    assert test_output[0][3] == org_lang_doc_index
-
-    # check if ignoring the langauge works
-    assert all(map(lambda x: x[1] != "en", test_output))
-
-
-@pytest.mark.parametrize(
-    "text, language, model_name",
-    [
-        (TEST_TEXT, "en", os.getenv("MODEL_NAME")),
-    ],
-)
-def test_create_embeddings(text, language, model_name):
-    openai_client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
-
-    n_sentences = len(sent_tokenize(text, langname(language).lower()))
-    print(f"Testing emebddings on {n_sentences} sentences")
-
-    start_single = time.time()
-    single_thread_result = create_embeddings(text, language, model_name, openai_client)
-    single_thread_time = time.time() - start_single
-
-    start_multi = time.time()
-    multi_thread_result = create_embeddings_multithreading(text, language, model_name, openai_client)
-    multi_thread_time = time.time() - start_multi
-
-    print(f"Single-threaded time: {single_thread_time:.4f}s")
-    print(f"Multi-threaded time: {multi_thread_time:.4f}s")
-    assert len(single_thread_result) == n_sentences
-    assert len(multi_thread_result) == n_sentences
-    assert len(single_thread_result[0]) == len(multi_thread_result[0])
-
-    # checks if the order is the same
-    for i, st_embedding in enumerate(single_thread_result):
-        the_same = cosine_similarity(st_embedding, multi_thread_result[i])
-        other = cosine_similarity(st_embedding, multi_thread_result[i - 1])
-        # the_same is the distance between embedding from single threading and multi threading at the same index
-        # this value should be very very very close to one (or rarely 1.0), like 0.9999999....
-        # other is the distance between single threading embedding and some other multi threading embedding,
-        # which should smaller value than the_same
-        print(the_same, other)
-        assert other < the_same
